chore: remove code dupe in trtllm/sglang/vllm backends
