+++ src/aiconfigurator/sdk/backends/base_backend.py
@@ -2,8 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import logging
-from abc import ABC, abstractmethod
 from collections import defaultdict
+from typing import ClassVar
 
 import numpy as np
 import pandas as pd
@@ -14,6 +14,8 @@ from aiconfigurator.sdk.inference_summary import InferenceSummary
 from aiconfigurator.sdk.models import BaseModel
 from aiconfigurator.sdk.perf_database import PerfDatabase
 from aiconfigurator.sdk.rust_engine_step import (
+    estimate_decode_step_latency_with_rust,
+    estimate_mixed_step_latency_with_rust,
     estimate_static_latency_breakdown_with_rust,
     should_use_rust_engine_step,
 )
@@ -21,22 +23,108 @@ from aiconfigurator.sdk.rust_engine_step import (
 logger = logging.getLogger(__name__)
 
 
-class BaseBackend(ABC):
-    """
-    Base class for all backends.
-    All backends should inherit from this class and implement the abstract methods.
-    All backends should implement the following methods:
+class BaseBackend:
+    """Base class for all inference backends.
 
-    Attributes:
+    Subclasses provide:
+        - ``self.name`` (set in ``__init__``).
+        - ``ACTIVATION_COEFFICIENTS``: per-model-family activation scaling factors.
+        - Optional overrides of memory-overhead constants
+          (``MIN_ACTIVATION_BYTES``, ``ACTIVATION_OVERHEAD_FRAC``, ``OTHERS_OVERHEAD_FRAC``)
+          and the agg-pipeline hooks (``_resolve_agg_kwargs``, ``_make_agg_cache_key``,
+          ``_memory_usage_kwargs_for_agg``, ``_oom_check_kwargs``, ``_moe_workspace_width``).
 
-    Methods:
-        run_static: this is common for all backends. It's implemented in this class.
-            If there might be some backend-specific logic, it should be implemented in the subclass.
-        run_agg: this is backend-specific. It should be implemented in the subclass.
-        find_best_agg_result_under_constraints: this is backend-specific.
-            It should be implemented in the subclass.
-        _get_memory_usage: this is backend-specific. It should be implemented in the subclass.
+    Concrete shared implementations:
+        - ``run_static`` / ``run_static_latency_only``: static-batching inference.
+        - ``run_agg``: continuous-batching inference for a single (b, ctx_tokens) point.
+        - ``find_best_agg_result_under_constraints``: SLA-constrained sweep over agg points.
+        - ``_get_memory_usage``: weights + activations + KV + nccl + others (model-family aware).
     """
+
+    # ---- Memory-model knobs (overridable by subclasses) ----------------
+    # Per-family activation scaling: family -> {tp_size: scalar}. The "default" key
+    # is used when a model_family is not in the table. Empty in BaseBackend; each
+    # subclass populates with its own table.
+    ACTIVATION_COEFFICIENTS: ClassVar[dict[str, dict[int, float]]] = {}
+
+    # Model families whose MoE block-scale dispatch workspace is added on top of
+    # the base activation budget.
+    MOE_WORKSPACE_FAMILIES: ClassVar[tuple[str, ...]] = (
+        "GEMMA4MOE",
+        "DEEPSEEK",
+        "DEEPSEEKV32",
+        "DEEPSEEKV4",
+        "KIMIK25",
+    )
+
+    # Minimum activation memory, in bytes (clamps from below).
+    MIN_ACTIVATION_BYTES: int = 70 * 1024 * 1024
+
+    # Multiplicative overhead applied after the base activation/others computation.
+    # SGLang sets these > 0 to model Python/runtime overhead.
+    ACTIVATION_OVERHEAD_FRAC: float = 0.0
+    OTHERS_OVERHEAD_FRAC: float = 0.0
+
+    def __init__(self):
+        # Flat dict keyed by tuple from ``_make_agg_cache_key``.
+        self._agg_cache: dict = {}
+        # Subclasses set the canonical name.
+        self.name = None
+
+    # ============== HOOKS (overridable by subclasses) ==================
+
+    def _moe_workspace_width(self, model: BaseModel, model_family: str, h: int) -> int:
+        """Feature width per token for MoE block-scale dispatch workspace.
+
+        Default: model's residual hidden size (``_hidden_size``), which equals
+        ``num_heads*head_size`` for most models but is wider for DeepSeek-V4's
+        attention expansion. TRT-LLM overrides this to use the raw ``h`` for
+        the DEEPSEEK family (legacy accounting, predates V4).
+        """
+        return getattr(model, "_hidden_size", h)
+
+    def _resolve_agg_kwargs(self, kwargs: dict, isl: int, osl: int) -> dict:
+        """Resolve backend-specific run_agg kwargs to defaults.
+
+        Default: returns an empty dict. TRT-LLM resolves ``max_seq_len`` /
+        ``max_num_tokens`` / ``free_gpu_memory_fraction`` here so both
+        ``run_agg`` and ``find_best_agg_result_under_constraints`` see the
+        same values when forwarding. Idempotent — calling with already-resolved
+        kwargs returns the same values.
+        """
+        return {}
+
+    def _make_agg_cache_key(
+        self,
+        isl: int,
+        osl: int,
+        b: int,
+        ctx_tokens: int,
+        engine_step_backend_key: str,
+        agg_extra: dict,
+    ) -> tuple:
+        """Build the cache key for ``run_agg`` results."""
+        return (isl, osl, b, ctx_tokens, engine_step_backend_key)
+
+    def _memory_usage_kwargs_for_agg(self, num_tokens: int, agg_extra: dict) -> dict:
+        """Kwargs for the ``_get_memory_usage`` call from ``run_agg``.
+
+        Default: pass the locally-computed ``num_tokens``. TRT-LLM passes
+        ``max_num_tokens`` (BuildConfig.max_num_tokens) for activation sizing
+        and forwards ``max_seq_len`` for KV cache sizing.
+        """
+        return {"num_tokens": num_tokens}
+
+    def _oom_check_kwargs(self, agg_extra: dict) -> dict:
+        """Extra kwargs for ``InferenceSummary.set_memory_and_check_oom``.
+
+        Default: none. TRT-LLM passes ``free_gpu_memory_fraction``,
+        ``kv_cache_reserved_fraction``, and ``kv_cache_tolerance`` to enable
+        the KV-cache capacity OOM check.
+        """
+        return {}
+
+    # ============== STATIC INFERENCE (shared) ==========================
 
     def _run_context_phase(
         self,
@@ -484,25 +572,543 @@ class BaseBackend(ABC):
         ctx_tokens_list.sort()
         return ctx_tokens_list
 
-    @abstractmethod
+    # ============== AGG STEP LATENCY HELPERS (shared) ==================
+
+    def _get_mix_step_latency(
+        self,
+        model: BaseModel,
+        database: PerfDatabase,
+        runtime_config: RuntimeConfig,
+        ctx_tokens: int,
+        gen_tokens: int,
+        isl: int,
+        osl: int,
+        prefix: int,
+    ) -> tuple[float, float, dict, dict]:
+        """Latency / energy for one mixed (chunked-prefill + decode) step.
+
+        Returns ``(latency_ms, energy_wms, per_op_latency, per_op_source)``.
+        The per-op dicts represent the breakdown for this step alone; the
+        caller stores them under the "mix_step" key of the run_agg
+        per-ops summary.
+        """
+        if should_use_rust_engine_step(runtime_config):
+            latency_ms = estimate_mixed_step_latency_with_rust(
+                model,
+                database,
+                ctx_tokens=ctx_tokens,
+                gen_tokens=gen_tokens,
+                isl=isl,
+                osl=osl,
+                prefix=prefix,
+            )
+            return (
+                latency_ms,
+                0.0,
+                {"rust_engine_step_mixed": latency_ms},
+                {"rust_engine_step_mixed": "rust"},
+            )
+
+        ctx_scale = runtime_config.seq_imbalance_correction_scale
+        gen_scale = runtime_config.gen_seq_imbalance_correction_scale
+
+        # Pass 1: combined single-batch inference to extract non-attention latency.
+        num_tokens_combined = ctx_tokens + gen_tokens
+        summary = self.run_static(
+            model,
+            database,
+            # num tokens for gemm needs to be adjusted for prefix, depends on the avg prefix len per request
+            RuntimeConfig(
+                batch_size=1,
+                beam_width=1,
+                isl=num_tokens_combined,
+                osl=1,
+                prefix=prefix * np.floor(ctx_tokens / isl),
+                seq_imbalance_correction_scale=ctx_scale,
+            ),
+            mode="static_ctx",
+        )
+        latency_dict = summary.get_context_latency_dict()
+        energy_wms_dict = summary.get_context_energy_wms_dict()
+        source_dict = summary.get_context_source_dict()
+        non_attention_latency_ms = 0.0
+        non_attention_energy_wms = 0.0
+        mix_non_attn_ops: dict[str, float] = {}
+        mix_non_attn_sources: dict[str, str] = {}
+        for layer_name, latency in latency_dict.items():
+            if layer_name != "context_attention":
+                non_attention_latency_ms += latency
+                non_attention_energy_wms += energy_wms_dict.get(layer_name, 0.0)
+                mix_non_attn_ops[layer_name] = latency
+                mix_non_attn_sources[layer_name] = source_dict.get(layer_name, "silicon")
+
+        # Pass 2: context attention split full isl over num_steps and averaged.
+        batch_size = np.ceil(ctx_tokens / isl)
+        summary = self.run_static(
+            model,
+            database,
+            RuntimeConfig(
+                batch_size=batch_size,
+                beam_width=1,
+                isl=isl,
+                osl=1,
+                prefix=prefix,
+                seq_imbalance_correction_scale=ctx_scale,
+            ),
+            mode="static_ctx",
+        )
+        latency_dict = summary.get_context_latency_dict()
+        energy_wms_dict = summary.get_context_energy_wms_dict()
+        ctx_attn_source = summary.get_context_source_dict().get("context_attention", "silicon")
+        scale_factor = np.ceil(isl / ctx_tokens)
+        ctx_attention_latency_ms = latency_dict["context_attention"] / scale_factor
+        ctx_attention_energy_wms = energy_wms_dict.get("context_attention", 0.0) / scale_factor
+
+        # Pass 3: generation attention (use isl + osl//2 for the avg seq len).
+        gen_attention_latency_ms = 0.0
+        gen_attention_energy_wms = 0.0
+        gen_attn_source = "silicon"
+        if gen_tokens > 0:
+            summary = self.run_static(
+                model,
+                database,
+                RuntimeConfig(
+                    batch_size=gen_tokens,
+                    beam_width=1,
+                    isl=isl + osl // 2,
+                    osl=2,
+                    gen_seq_imbalance_correction_scale=gen_scale,
+                ),
+                mode="static_gen",
+            )
+            latency_dict = summary.get_generation_latency_dict()
+            energy_wms_dict = summary.get_generation_energy_wms_dict()
+            gen_attention_latency_ms = latency_dict["generation_attention"]
+            gen_attention_energy_wms = energy_wms_dict.get("generation_attention", 0.0)
+            gen_attn_source = summary.get_generation_source_dict().get("generation_attention", "silicon")
+
+        per_ops_step_data = {
+            **mix_non_attn_ops,
+            "context_attention (scaled)": ctx_attention_latency_ms,
+            "generation_attention": gen_attention_latency_ms,
+        }
+        per_ops_step_source = {
+            **mix_non_attn_sources,
+            "context_attention (scaled)": ctx_attn_source,
+            "generation_attention": gen_attn_source,
+        }
+
+        total_latency_ms = non_attention_latency_ms + ctx_attention_latency_ms + gen_attention_latency_ms
+        total_energy_wms = non_attention_energy_wms + ctx_attention_energy_wms + gen_attention_energy_wms
+        return total_latency_ms, total_energy_wms, per_ops_step_data, per_ops_step_source
+
+    def _get_genonly_step_latency(
+        self,
+        model: BaseModel,
+        database: PerfDatabase,
+        runtime_config: RuntimeConfig,
+        gen_tokens: int,
+        isl: int,
+        osl: int,
+    ) -> tuple[float, float, dict, dict]:
+        """Latency / energy for one generation-only step.
+
+        Returns ``(latency_ms, energy_wms, per_op_latency, per_op_source)``.
+        When ``gen_tokens <= 0`` both totals are 0 and the per-op dicts are empty.
+        """
+        if gen_tokens <= 0:
+            return 0.0, 0.0, {}, {}
+        if should_use_rust_engine_step(runtime_config):
+            latency_ms = estimate_decode_step_latency_with_rust(
+                model,
+                database,
+                gen_tokens=gen_tokens,
+                isl=isl,
+                osl=osl,
+            )
+            return (
+                latency_ms,
+                0.0,
+                {"rust_engine_step_generation": latency_ms},
+                {"rust_engine_step_generation": "rust"},
+            )
+
+        gen_scale = runtime_config.gen_seq_imbalance_correction_scale
+        summary = self.run_static(
+            model,
+            database,
+            RuntimeConfig(
+                batch_size=gen_tokens,
+                beam_width=1,
+                isl=isl + osl // 2,
+                osl=2,
+                gen_seq_imbalance_correction_scale=gen_scale,
+            ),
+            mode="static_gen",
+        )
+        latency_dict = summary.get_generation_latency_dict()
+        energy_wms_dict = summary.get_generation_energy_wms_dict()
+        source_dict = summary.get_generation_source_dict()
+        total_latency_ms = 0.0
+        total_energy_wms = 0.0
+        per_ops_step_data: dict[str, float] = {}
+        per_ops_step_source: dict[str, str] = {}
+        for layer_name, latency in latency_dict.items():
+            total_latency_ms += latency
+            total_energy_wms += energy_wms_dict.get(layer_name, 0.0)
+            per_ops_step_data[layer_name] = latency
+            per_ops_step_source[layer_name] = source_dict.get(layer_name, "silicon")
+        return total_latency_ms, total_energy_wms, per_ops_step_data, per_ops_step_source
+
+    # ============== AGG INFERENCE (shared) =============================
+
     def run_agg(
         self, model: BaseModel, database: PerfDatabase, runtime_config: RuntimeConfig, **kwargs
     ) -> InferenceSummary:
-        """
-        Run the agg inference.
-        """
-        pass
+        """Run the agg (continuous-batching) inference for a single (b, ctx_tokens) point."""
+        isl = runtime_config.isl
+        osl = runtime_config.osl
+        prefix = runtime_config.prefix
+        b = runtime_config.batch_size
+        engine_step_backend_key = "rust" if should_use_rust_engine_step(runtime_config) else "python"
+        ctx_tokens = kwargs.get("ctx_tokens")
+        assert ctx_tokens is not None, "ctx_tokens is required"
+        balance_score = isl * b / ctx_tokens / osl
 
-    @abstractmethod
+        # Backend-specific kwargs (TRT-LLM: max_seq_len / max_num_tokens /
+        # free_gpu_memory_fraction; others: {}).
+        agg_extra = self._resolve_agg_kwargs(kwargs, isl=isl, osl=osl)
+
+        cache_key = self._make_agg_cache_key(isl, osl, b, ctx_tokens, engine_step_backend_key, agg_extra)
+        cached = self._agg_cache.get(cache_key)
+        if cached is not None:
+            return cached
+
+        # Compute num_mix_steps / num_genonly_steps within osl steps such that
+        # all ctx tokens are consumed.
+        steps_to_finish_ctx = np.ceil(isl * b / ctx_tokens)
+        num_mix_steps = num_genonly_steps = 0
+        num_mix_steps_for_tpot_calc = 0  # correction for tpot calc only
+        if b > 1:
+            if steps_to_finish_ctx >= osl:
+                num_mix_steps = steps_to_finish_ctx
+                num_mix_ctx_tokens = ctx_tokens
+                num_mix_gen_tokens = max(1, b // (steps_to_finish_ctx / osl))
+                num_genonly_steps = 0
+                num_genonly_tokens = 0
+                num_mix_steps_for_tpot_calc = num_mix_steps
+            else:
+                # 3-step is an empirical correction for pipelining requests where new requests
+                # cannot be enqueued immediately after last request's exit
+                num_mix_steps = steps_to_finish_ctx
+                num_mix_ctx_tokens = ctx_tokens
+                num_mix_gen_tokens = b - np.ceil(ctx_tokens / isl)  # the error check is outside
+                assert num_mix_gen_tokens >= 1, (
+                    f"num_mix_gen_tokens: {num_mix_gen_tokens}, b: {b}, ctx_tokens: {ctx_tokens}, isl: {isl}"
+                )
+                num_genonly_steps = osl - num_mix_steps
+                num_genonly_tokens = b
+                num_mix_steps_for_tpot_calc = max(1, num_mix_steps - 3)
+        elif b == 1:
+            # special case for b=1
+            num_mix_steps = 1
+            num_mix_ctx_tokens = ctx_tokens
+            num_mix_gen_tokens = 0
+            num_genonly_steps = osl - 1
+            num_genonly_tokens = 1
+            num_mix_steps_for_tpot_calc = 0
+
+        # Step-latency helpers (return (latency_ms, energy_wms, per_op_data, per_op_source)).
+        per_ops_data: dict[str, dict] = {}
+        per_ops_source: dict[str, dict] = {}
+
+        mix_step_latency_ms, mix_step_energy_wms, mix_per_ops, mix_per_ops_src = self._get_mix_step_latency(
+            model, database, runtime_config, num_mix_ctx_tokens, num_mix_gen_tokens, isl, osl, prefix
+        )
+        per_ops_data["mix_step"] = mix_per_ops
+        per_ops_source["mix_step"] = mix_per_ops_src
+
+        (
+            genonly_step_latency_ms,
+            genonly_step_energy_wms,
+            genonly_per_ops,
+            genonly_per_ops_src,
+        ) = self._get_genonly_step_latency(model, database, runtime_config, num_genonly_tokens, isl, osl)
+        if genonly_per_ops:
+            per_ops_data["genonly_step"] = genonly_per_ops
+            per_ops_source["genonly_step"] = genonly_per_ops_src
+
+        # TTFT: assume a 10x request queue, capped correction factor at 4.
+        ttft = mix_step_latency_ms * np.ceil(isl / ctx_tokens)
+        correction_factor = min(2 + (steps_to_finish_ctx - 3) / 2 / 10, 4)
+        ttft *= correction_factor
+        logger.debug(
+            f"ttft correction factor: {2 + (steps_to_finish_ctx - 3) / 2 / 10} capped to "
+            f"{correction_factor} when b: {b}, ctx_tokens: {ctx_tokens} isl {isl}"
+        )
+
+        tpot = (mix_step_latency_ms * num_mix_steps_for_tpot_calc + genonly_step_latency_ms * num_genonly_steps) / (
+            num_mix_steps_for_tpot_calc + num_genonly_steps
+        )
+        output_throughput = (
+            1000 / (num_mix_steps * mix_step_latency_ms + num_genonly_steps * genonly_step_latency_ms) * b * (osl - 1)
+        )
+        logger.debug(
+            f"ctx_tokens: {ctx_tokens}, b: {b}, osl: {osl}, isl: {isl}, "
+            f"num_mix_steps: {num_mix_steps}, num_genonly_steps: {num_genonly_steps}, "
+            f"num_mix_ctx_tokens: {num_mix_ctx_tokens}, "
+            f"num_mix_gen_tokens: {num_mix_gen_tokens}, "
+            f"num_genonly_tokens: {num_genonly_tokens}"
+        )
+        logger.debug(f"mix_step_latency: {mix_step_latency_ms} ms, genonly_step_latency: {genonly_step_latency_ms} ms")
+        logger.debug(
+            f"mix_step_energy: {mix_step_energy_wms} W·ms, genonly_step_energy: {genonly_step_energy_wms} W·ms"
+        )
+        logger.debug(f"ttft: {ttft}, tpot: {tpot}, output_throughput: {output_throughput}")
+
+        # Weighted average power: total energy / total latency.
+        total_energy_wms = num_mix_steps * mix_step_energy_wms + num_genonly_steps * genonly_step_energy_wms
+        total_latency_ms = num_mix_steps * mix_step_latency_ms + num_genonly_steps * genonly_step_latency_ms
+        agg_power_avg_w = total_energy_wms / total_latency_ms if total_latency_ms > 0 else 0.0
+        logger.debug(f"Aggregated power: {agg_power_avg_w}W (from {total_energy_wms} W·ms / {total_latency_ms} ms)")
+
+        num_ctx_requests = np.ceil(ctx_tokens / isl)
+        num_gen_requests = b - num_ctx_requests
+        if b == 1:
+            num_ctx_requests = 1
+            num_gen_requests = 1
+
+        # correct output_throughput and concurrency for attention dp (global batch)
+        scale_factor = model.config.pp_size * model.config.attention_dp_size
+        output_throughput = output_throughput * scale_factor
+        concurrency = b * scale_factor
+
+        request_rate = output_throughput / (osl - 1)
+        if b > 1:
+            # will not be corrected by balance score when it's larger than 1.0
+            # in order to indicate what's happening
+            num_tokens = num_gen_requests + ctx_tokens
+        else:
+            num_tokens = ctx_tokens
+
+        memory = self._get_memory_usage(
+            model,
+            database,
+            b,
+            1,
+            isl,
+            osl,
+            prefix=prefix,
+            **self._memory_usage_kwargs_for_agg(num_tokens=num_tokens, agg_extra=agg_extra),
+        )
+        tp = model.config.tp_size
+        pp = model.config.pp_size
+        dp = model.config.attention_dp_size
+        moe_tp = model.config.moe_tp_size
+        moe_ep = model.config.moe_ep_size
+        tokens_s_gpu = output_throughput / pp / tp / dp
+        tokens_s_user = 1000 / tpot
+        seq_s = request_rate
+        seq_s_gpu = seq_s / pp / tp / dp
+        tokens_s = output_throughput
+        request_latency = ttft + tpot * max(osl - 1, 0)
+        num_total_gpus = tp * pp * dp
+        parallel = f"tp{tp}pp{pp}dp{dp}etp{moe_tp}ep{moe_ep}"
+        gemm = model.config.gemm_quant_mode.name
+        kvcache = model.config.kvcache_quant_mode.name
+        fmha = model.config.fmha_quant_mode.name
+        moe = model.config.moe_quant_mode.name
+        comm = model.config.comm_quant_mode.name
+        mem = memory["total"]
+
+        result_dict = {
+            "model": model.model_path,
+            "isl": isl,
+            "osl": osl,
+            "prefix": prefix,
+            "concurrency": concurrency,
+            "request_rate": request_rate,
+            "bs": b,
+            "global_bs": b * model.config.attention_dp_size,
+            "ttft": ttft,
+            "tpot": tpot,
+            "seq/s": seq_s,
+            "seq/s/gpu": seq_s_gpu,
+            "tokens/s": tokens_s,
+            "tokens/s/gpu": tokens_s_gpu,
+            "tokens/s/user": tokens_s_user,
+            "request_latency": request_latency,
+            "num_total_gpus": num_total_gpus,
+            "tp": tp,
+            "pp": pp,
+            "dp": dp,
+            "moe_tp": moe_tp,
+            "moe_ep": moe_ep,
+            "parallel": parallel,
+            "gemm": gemm,
+            "kvcache": kvcache,
+            "fmha": fmha,
+            "moe": moe,
+            "comm": comm,
+            "memory": mem,
+            "balance_score": balance_score,
+            "num_ctx_reqs": num_ctx_requests,
+            "num_gen_reqs": num_gen_requests,
+            "num_tokens": num_tokens,
+            "ctx_tokens": ctx_tokens,
+            "gen_tokens": num_gen_requests,
+            "backend": database.backend,
+            "version": database.version,
+            "system": database.system,
+            "power_w": agg_power_avg_w,
+        }
+        result = pd.DataFrame([result_dict], columns=common.ColumnsAgg).round(3)
+        summary = InferenceSummary(RuntimeConfig(isl=isl, osl=osl))
+        summary.set_memory_and_check_oom(
+            memory,
+            database.system_spec["gpu"]["mem_capacity"],
+            **self._oom_check_kwargs(agg_extra),
+        )
+        summary.set_summary_df(result)
+        summary.set_result_dict(result_dict)
+
+        # Scheduling counters: aggregate sums, not DB queries — recorded in
+        # per_ops_data only; no per-op source applies.
+        per_ops_data["scheduling"] = {
+            "num_mix_steps": float(num_mix_steps),
+            "num_genonly_steps": float(num_genonly_steps),
+            "mix_step_latency_ms": float(mix_step_latency_ms),
+            "genonly_step_latency_ms": float(genonly_step_latency_ms),
+        }
+        summary.set_per_ops_data(per_ops_data)
+        summary.set_per_ops_source(per_ops_source)
+
+        self._agg_cache[cache_key] = summary
+        return summary
+
     def find_best_agg_result_under_constraints(
         self, model: BaseModel, database: PerfDatabase, runtime_config: RuntimeConfig, **kwargs
     ) -> InferenceSummary:
         """
         Find the best agg result under constraints.
-        """
-        pass
 
-    @abstractmethod
+        Args:
+            model: the model to be tested
+            database: the database to be tested
+            runtime_config: the runtime configuration
+            top_k: the number of best results to return
+            max_batch_size: the maximum batch size to test
+            ctx_stride: the stride of ctx tokens to test, it will impact the time to run the test.
+            enable_chunked_prefill: whether to enable chunked prefill, it will impact the time to
+                run the test while have little impact on the result. Default off.
+            **kwargs: additional backend-specific kwargs (e.g. TRT-LLM accepts
+                ``max_seq_len`` and ``free_gpu_memory_fraction``).
+
+        Returns:
+            A summary of the best agg result under constraints.
+        """
+        isl = runtime_config.isl
+        osl = runtime_config.osl
+        ttft = runtime_config.ttft
+        tpot = runtime_config.tpot
+        prefix = runtime_config.prefix
+        top_k = kwargs.get("top_k", 1)
+        max_batch_size = kwargs.get("max_batch_size", 512)
+        ctx_stride = kwargs.get("ctx_stride", 512)
+        enable_chunked_prefill = kwargs.get("enable_chunked_prefill", False)
+
+        # Resolve backend-specific kwargs once; forward into run_agg so each
+        # (b, ctx_tokens) point sees the same backend params.
+        sweep_extra = self._resolve_agg_kwargs(kwargs, isl=isl, osl=osl)
+
+        # when b is larger than 1024, the result is not good as the data collection is not enough
+        # to cover this.
+        b_list_default = (
+            list(range(1, 16, 1))
+            + list(range(16, 32, 4))
+            + list(range(32, 64, 8))
+            + list(range(64, 256, 16))
+            + list(range(256, 512, 32))
+            + list(range(512, 1024, 256))
+            + [1024]
+        )
+
+        # sweep for batch_size and ctx_tokens
+        # ctx_tokens will have a step of ctx_stride. When it's larger than 8192, we will increase
+        # the step to ctx_stride_large.
+        # outer_loop is over batch_size dimention, from 1 to max_batch_size
+        # inner_loop is over ctx_tokens dimention, from 0 to max_ctx_tokens where it's
+        # max(8192, 4*isl).
+        # during the loop, as b, ctx_tokens and system memory are monotonic, we can break the
+        # inner loop when the system is oom.
+        b_list = [b for b in b_list_default if b <= max_batch_size]
+        ctx_tokens_list = self._get_ctx_tokens_list_for_agg_sweep(isl, ctx_stride, enable_chunked_prefill)
+
+        results_df = pd.DataFrame(columns=common.ColumnsAgg)
+        results_dict_list: list[dict] = []
+        results_per_ops_source: list[dict | None] = []  # aligned with results_dict_list
+        capped_b: list[int] = []
+        all_oom = True
+        for b in b_list:
+            for ctx_tokens in ctx_tokens_list:
+                if b - np.ceil(ctx_tokens / isl) < 0:  # allow b==1
+                    break
+
+                if b > 1 and (
+                    b - np.ceil(ctx_tokens / isl) < 1
+                ):  # general case, to ensure there's at least one gen req
+                    break
+
+                # filter out repeated records for balance score correction
+                balance_score = isl * b / ctx_tokens / osl
+                if balance_score > 1:
+                    gen_tokens = b // balance_score
+                    if gen_tokens > 1 and gen_tokens in capped_b:
+                        continue
+                    else:
+                        capped_b.append(gen_tokens)
+
+                summary = self.run_agg(
+                    model=model,
+                    database=database,
+                    runtime_config=RuntimeConfig(
+                        batch_size=b,
+                        isl=isl,
+                        osl=osl,
+                        prefix=prefix,
+                        seq_imbalance_correction_scale=runtime_config.seq_imbalance_correction_scale,
+                        engine_step_backend=runtime_config.engine_step_backend,
+                    ),
+                    ctx_tokens=ctx_tokens,
+                    **sweep_extra,
+                )
+
+                if summary.check_oom() or summary.check_kv_cache_oom():
+                    break  # larger ctx tokens will cause oom
+                all_oom = False
+                result_dict = summary.get_result_dict()
+                if result_dict and result_dict["tpot"] <= tpot and result_dict["ttft"] <= ttft:
+                    results_dict_list.append(result_dict)
+                    results_per_ops_source.append(summary.get_per_ops_source())
+
+        if results_dict_list:
+            results_df = pd.DataFrame(results_dict_list, columns=common.ColumnsAgg).round(3)
+            # Carry per-row per_ops_source as an object column, sorted/truncated alongside the
+            # standard columns. report_and_save.py strips this before writing best_config_topn.csv
+            # and emits one per_ops_source.json per topN/ subdir.
+            results_df["_per_ops_source"] = results_per_ops_source
+
+        sorted_results_df = results_df.sort_values(by="seq/s", ascending=False).round(3)
+        if top_k > 0:
+            sorted_results_df = sorted_results_df.head(top_k)
+
+        summary = InferenceSummary(runtime_config)
+        summary.set_summary_df(sorted_results_df)
+        summary.set_oom(all_oom)
+        return summary
+
+    # ============== MEMORY USAGE (shared) ==============================
+
     def _get_memory_usage(
         self,
         model: BaseModel,
@@ -513,6 +1119,7 @@ class BaseBackend(ABC):
         osl: int,
         num_tokens: int = 0,
         prefix: int = 0,
+        max_seq_len: int | None = None,
     ) -> dict[str, float]:
         """
         Get the memory usage of the backend.
@@ -520,5 +1127,67 @@ class BaseBackend(ABC):
         Args:
             prefix: number of prefix tokens (part of isl) whose KV is already cached
                 (per-request) and does not need activation computation.
+            max_seq_len: per-slot KV cache pre-allocation budget. Defaults to
+                ``isl + beam_width * osl`` when not supplied.
         """
-        pass
+        weights = 0.0
+        for op in model.context_ops:
+            weights += op.get_weights()
+        # count weights on a single GPU
+        weights /= model.config.pp_size
+
+        h = model._num_heads * model._head_size
+        if num_tokens == 0:
+            num_tokens = (isl - prefix) * batch_size
+
+        tp_clamped = min(model.config.tp_size, 8)
+        family = model.model_family
+        coeffs_table = self.ACTIVATION_COEFFICIENTS
+        coeffs = coeffs_table.get(family, coeffs_table.get("default", {1: 10, 2: 6, 4: 5, 8: 5}))
+        activations = 2 * num_tokens * h * coeffs[tp_clamped]
+
+        # MoE block-scale dispatch workspace (only for families that pay this cost).
+        # 128 = block scale; 4 = float bytes.
+        if family in self.MOE_WORKSPACE_FAMILIES:
+            moe_h = self._moe_workspace_width(model, family, h)
+            activations += (
+                num_tokens
+                * moe_h
+                * model.config.attention_dp_size
+                * model._num_experts
+                * model._topk
+                / model.config.moe_ep_size
+                / 128
+                * 4
+            )
+
+        activations = max(activations, self.MIN_ACTIVATION_BYTES)
+
+        # MTP correction: additional activation memory for draft tokens (applies to all models)
+        if model.config.nextn > 0:
+            activations = activations * (model.config.nextn + 1)
+
+        # Backend-level activation overhead (SGLang only by default).
+        if self.ACTIVATION_OVERHEAD_FRAC > 0:
+            activations *= 1.0 + self.ACTIVATION_OVERHEAD_FRAC
+
+        seq_tokens = max_seq_len if max_seq_len is not None else isl + beam_width * osl
+        kvcache = batch_size * model.get_kvcache_bytes_per_sequence(seq_tokens)
+        # should not be divided by pp_size as you need to hold all kvcache for stages.
+
+        # starting from 2.22
+        nccl_mem = database.system_spec["misc"]["nccl_mem"][tp_clamped]
+        # cuda, cublas, etc.
+        others_mem = database.system_spec["misc"]["other_mem"]
+        if self.OTHERS_OVERHEAD_FRAC > 0:
+            others_mem *= 1.0 + self.OTHERS_OVERHEAD_FRAC
+
+        one_gib = 1 << 30
+        return {
+            "total": (weights + activations + kvcache + nccl_mem + others_mem) / one_gib,
+            "weights": weights / one_gib,
+            "activations": activations / one_gib,
+            "kvcache": kvcache / one_gib,
+            "nccl": nccl_mem / one_gib,
+            "others": others_mem / one_gib,
+        }
